add fragment_rocket content

+++ app/src/main/java/com/x5bartsoft/spacex/model/response/launchdetail/FirstStage.kt
@@ -13,7 +13,7 @@ data class FirstStage(
     @SerializedName("engines")
     val engines: Int,
     @SerializedName("fuel_amount_tons")
-    val fuelAmountTons: Int,
+    val fuelAmountTons: Double,
     @SerializedName("reusable")
     val reusable: Boolean,
     @SerializedName("thrust_sea_level")
